First implementation to display multilingual email parts (RFC8255)
Written during the #IETFHackathon [1] of the #IETF101 [2].
[1] : https://trac.ietf.org/trac/ietf/meeting/wiki/101hackathon
[2] : https://www.ietf.org/how/meetings/101/

This code is **not stable** nor **correct** regarding to the neomutt code style / contribution guide.
This just demonstrated the Proof Of Concept to display in a MUA email build with RFC8255.
**Use at your own risk.**

A RFC8255 reflector has also be developped during the #IETFHackathon, see here : https://github.com/igit/rfc8255-reflector